splitting parts of Footer into separate client-side components
